feat: base element for datepicker buttons
