+++ src/components/datepicker/datepicker/datepicker.ts
@@ -11,8 +11,7 @@ import { EventEmitter } from '../../core/eventing';
 import { i18nDateChangedTo, i18nDatePickerPlaceholder } from '../../core/i18n';
 import type { SbbDateLike, SbbValidationChangeEvent } from '../../core/interfaces';
 import { AgnosticMutationObserver } from '../../core/observers';
-import type { SbbDatepickerNextDayElement } from '../datepicker-next-day';
-import type { SbbDatepickerPreviousDayElement } from '../datepicker-previous-day';
+import type { SbbDatepickerButton } from '../common/datepicker-button';
 import type { SbbDatepickerToggleElement } from '../datepicker-toggle';
 
 import style from './datepicker.scss?lit&inline';
@@ -34,10 +33,7 @@ export interface SbbInputUpdateEvent {
  * @param trigger The id or the reference of the SbbDatePicker.
  */
 export function getDatePicker(
-  element:
-    | SbbDatepickerPreviousDayElement
-    | SbbDatepickerNextDayElement
-    | SbbDatepickerToggleElement,
+  element: SbbDatepickerButton | SbbDatepickerToggleElement,
   trigger?: string | HTMLElement,
 ): SbbDatepickerElement | null | undefined {
   if (!trigger) {
