+++ src/components/datepicker/index.ts
@@ -2,3 +2,4 @@ export * from './datepicker';
 export * from './datepicker-next-day';
 export * from './datepicker-previous-day';
 export * from './datepicker-toggle';
+export * from './common';
